refactor: move component-specific constants to OnboardingScreen and exporting them from there
