+++ app/screens/OnboardingScreen/OnboardingScreen.tsx
@@ -14,12 +14,49 @@ import {
   OnboardingBackground,
 } from '@components';
 
-import {
-  ONBOARDING_BACKGROUND_COLORS,
-  ONBOARDING_TITLES,
-} from '@config/constants';
-
 import { stylesheet } from './styles';
+
+/**
+ * Total number of steps in the onboarding process.
+ */
+export const ONBOARDING_TOTAL_STEPS = 3;
+
+/**
+ * Array of translation keys for onboarding screen titles.
+ */
+export const ONBOARDING_TITLES = [
+  'onboardingTitle1',
+  'onboardingTitle2',
+  'onboardingTitle3',
+];
+
+/**
+ * Array of colors used for transitions between onboarding screens.
+ */
+export const ONBOARDING_TRANSITION_COLORS = ['#FFF2D9', '#FFF2D9', '#F7B327'];
+
+/**
+ * Array of background colors used for each onboarding screen.
+ */
+export const ONBOARDING_BACKGROUND_COLORS = ['#FFF2D9', '#F7B327', '#FFF2D9'];
+
+/**
+ * Array of images used on each onboarding screen.
+ */
+export const ONBOARDING_IMAGES = [
+  { image: require('../../resources/assets/images/onboarding-cat-1.png') },
+  { image: require('../../resources/assets/images/onboarding-cat-2.png') },
+  { image: require('../../resources/assets/images/onboarding-cat-3.png') },
+];
+
+/**
+ * Array of images used as background on each onboarding screen.
+ */
+export const ONBOARDING_BACKGROUND_IMAGES = [
+  { image: require('../../resources/assets/images/onboarding-background-1.png') },
+  { image: require('../../resources/assets/images/onboarding-background-2.png') },
+  { image: require('../../resources/assets/images/onboarding-background-1.png') },
+];
 
 export const OnboardingScreen = () => {
   const [step, setStep] = useState<number>(1);
